add header UI, update base folder
update constructor folder
