+++ package.json
@@ -15,10 +15,13 @@
         "web-vitals": "^2.1.4"
     },
     "devDependencies": {
+        "autoprefixer": "^10.4.12",
         "babel-plugin-module-resolver": "^4.1.0",
         "customize-cra": "^1.0.0",
+        "postcss": "^8.4.18",
         "react-app-rewired": "^2.2.1",
-        "sass": "^1.53.0"
+        "sass": "^1.53.0",
+        "tailwindcss": "^3.1.8"
     },
     "scripts": {
         "start": "react-app-rewired start",
