+++ src/assets/images/index.js
@@ -1,6 +1,6 @@
 const images = {
-    logo: require('~assets/images/logo.png'),
-    noImage: require('~assets/images/no-image.png'),
+    logo: require('~/assets/images/logo.png'),
+    noImage: require('~/assets/images/no-image.png'),
 };
 
 export default images;
