+++ src/base.scss
@@ -13,7 +13,6 @@
 }
 
 html {
-    font-size: 62.5%;
 }
 
 body {
