--- src/components/Layout/components/Header/index.js
@@ -1,5 +1,0 @@
-function Header() {
-    return <h1>Header</h1>;
-}
-
-export default Header;
