--- src/components/Layout/index.js
@@ -1,1 +1,0 @@
-export { default as DefaultLayout } from './DefaultLayout';
