+++ src/index.css
@@ -1,0 +1,8 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+.container {
+    @apply w-full mx-auto px-4 2xl:max-w-screen-2xl 2xl:px-32 
+    xl:max-w-screen-xl xl:px-10 lg:max-w-screen-lg md:max-w-screen-md sm:max-w-screen-sm;
+}
