+++ src/layouts/DefaultLayout/DefaultLayout.jsx
@@ -1,9 +1,9 @@
-import Header from '~/components/Layout/components/Header';
+import Header from '~/layouts/components/Header';
 function DefaultLayout({ children }) {
     return (
         <div>
             <Header />
-            <div className="container">{children}</div>
+            <div>{children}</div>
         </div>
     );
 }
